# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API for ChallengePipefyIntegration",
        "title": "ChallengePipefyIntegration API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/clientes": {
            "post": {
                "description": "Creates a client and maps it to a Pipefy card",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "clients"
                ],
                "summary": "Create a new client",
                "parameters": [
                    {
                        "description": "Client data",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.CreateClientRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/pkg.Response-models_ClientResponse"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/pkg.Response-models_ClientResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/pkg.Response-models_ClientResponse"
                        }
                    },
                    "404": {
                        "description": "Not found",
                        "schema": {
                            "$ref": "#/definitions/pkg.Response-models_ClientResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/pkg.Response-models_ClientResponse"
                        }
                    }
                }
            }
        },
        "/health": {
            "get": {
                "description": "Returns ok if the server is running",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "health"
                ],
                "summary": "Health check",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/webhooks/pipefy/card-updated": {
            "post": {
                "description": "Receives a webhook event from Pipefy, applies priority rules and updates the client status",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "webhooks"
                ],
                "summary": "Process Pipefy card update event",
                "parameters": [
                    {
                        "description": "Webhook event data",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.WebhookEventRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Processed",
                        "schema": {
                            "$ref": "#/definitions/pkg.Response-models_WebhookEventResponse"
                        }
                    },
                    "400": {
                        "description": "Bad request or duplicate event",
                        "schema": {
                            "$ref": "#/definitions/pkg.Response-models_WebhookEventResponse"
                        }
                    },
                    "404": {
                        "description": "Client not found",
                        "schema": {
                            "$ref": "#/definitions/pkg.Response-models_WebhookEventResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/pkg.Response-models_WebhookEventResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "models.ClientPriority": {
            "type": "string",
            "enum": [
                "prioridade_alta",
                "prioridade_normal"
            ],
            "x-enum-varnames": [
                "PriorityHigh",
                "PriorityNormal"
            ]
        },
        "models.ClientResponse": {
            "type": "object",
            "properties": {
                "cliente_email": {
                    "type": "string"
                },
                "cliente_nome": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "prioridade": {
                    "$ref": "#/definitions/models.ClientPriority"
                },
                "status": {
                    "$ref": "#/definitions/models.ClientStatus"
                },
                "tipo_solicitacao": {
                    "type": "string"
                },
                "valor_patrimonio": {
                    "type": "number"
                }
            }
        },
        "models.ClientStatus": {
            "type": "string",
            "enum": [
                "Aguardando Análise",
                "Processado"
            ],
            "x-enum-varnames": [
                "StatusPending",
                "StatusProcessed"
            ]
        },
        "models.CreateClientRequest": {
            "type": "object",
            "required": [
                "cliente_email",
                "cliente_nome",
                "tipo_solicitacao",
                "valor_patrimonio"
            ],
            "properties": {
                "cliente_email": {
                    "type": "string",
                    "maxLength": 255
                },
                "cliente_nome": {
                    "type": "string",
                    "maxLength": 255
                },
                "tipo_solicitacao": {
                    "type": "string",
                    "maxLength": 255
                },
                "valor_patrimonio": {
                    "type": "number",
                    "maximum": 9999999999999
                }
            }
        },
        "models.WebhookEventRequest": {
            "type": "object",
            "required": [
                "card_id",
                "cliente_email",
                "event_id",
                "timestamp"
            ],
            "properties": {
                "card_id": {
                    "type": "string",
                    "maxLength": 255
                },
                "cliente_email": {
                    "type": "string",
                    "maxLength": 255
                },
                "event_id": {
                    "type": "string",
                    "maxLength": 255
                },
                "timestamp": {
                    "type": "string"
                }
            }
        },
        "models.WebhookEventResponse": {
            "type": "object",
            "properties": {
                "cliente_email": {
                    "type": "string"
                },
                "event_id": {
                    "type": "string"
                },
                "prioridade": {
                    "$ref": "#/definitions/models.ClientPriority"
                },
                "processed_at": {
                    "type": "string"
                },
                "status": {
                    "$ref": "#/definitions/models.ClientStatus"
                }
            }
        },
        "pkg.Response-models_ClientResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/models.ClientResponse"
                },
                "error": {
                    "type": "string"
                }
            }
        },
        "pkg.Response-models_WebhookEventResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/models.WebhookEventResponse"
                },
                "error": {
                    "type": "string"
                }
            }
        }
    }
}